requirements.txt
# Core ML — Mac M5 (MLX path)
mlx-lm>=0.22.0

# Core ML — HuggingFace (MPS / Colab paths)
torch>=2.3.0
transformers>=4.50.0
peft>=0.14.0
trl>=0.12.0
bitsandbytes>=0.43.0
accelerate>=0.27.0
datasets>=2.20.0

# Parsers
pdfplumber>=0.10.0
python-docx>=1.1.0
html2text>=2024.2.26

# Anonimización
spacy>=3.7.0
# python -m spacy download es_core_news_sm

# Evaluación
nltk>=3.8.0
scikit-learn>=1.4.0
numpy>=1.26.0
tqdm>=4.66.0

# Backend
fastapi>=0.110.0
httpx>=0.27.0
uvicorn>=0.29.0

# RAG (opcional — MVP sin esto)
# chromadb>=0.4.0
# sentence-transformers>=2.7.0
